Fully move arrow_sql_gen
