+++ Cargo.toml
@@ -10,12 +10,19 @@ arrow = "52.0.0"
 bigdecimal = "0.4.5"
 bigdecimal_0_3_0 = { package = "bigdecimal", version = "0.3.0" }
 byteorder = "1.5.0"
+chrono = "0.4.38"
 fallible-iterator = "0.3.0"
-mysql_async = "0.34.1"
+mysql_async = { version = "0.34.1", features = ["native-tls-tls", "chrono"], optional = true }
+rusqlite = { version = "0.31.0", optional = true }
 sea-query = { version = "0.30.7", features = ["backend-sqlite", "backend-postgres", "postgres-array", "with-rust_decimal", "with-bigdecimal", "with-time"] }
 snafu = "0.8.3"
 time = "0.3.36"
-tokio = "1.38.0"
-tokio-postgres = { version = "0.7.10", features = ["with-chrono-0_4", "with-uuid-1"] }
+tokio = { version = "1.38.0", features = ["macros"] }
+tokio-postgres = { version = "0.7.10", features = ["with-chrono-0_4", "with-uuid-1"], optional = true }
 tracing = "0.1.40"
-uuid = "1.9.1"
+uuid = { version = "1.9.1", optional = true }
+
+[features]
+mysql = ["dep:mysql_async"]
+postgres = ["dep:tokio-postgres", "dep:uuid"]
+sqlite = ["dep:rusqlite"]
