+++ Makefile
@@ -1,6 +1,10 @@
 all:
-	cargo build
+	cargo build --all-features
 
 .PHONY: test
 test:
-	cargo test
+	cargo test --all-features
+
+.PHONY: lint
+lint:
+	cargo clippy --all-features
