+++ src/sql/arrow_sql_gen/mysql.rs
@@ -1,4 +1,4 @@
-use crate::arrow::map_data_type_to_array_builder_optional;
+use crate::sql::arrow_sql_gen::arrow::map_data_type_to_array_builder_optional;
 use arrow::{
     array::{
         ArrayBuilder, ArrayRef, Date32Builder, Decimal128Builder, Float32Builder, Float64Builder,
